get rid of SET_PED_RANDOM_COMPONENT_VARIATION
