+++ src/views/settings/view_outfit_editor.cpp
@@ -1,50 +1,18 @@
 #include "natives.hpp"
 #include "pointers.hpp"
+#include "util/outfit.hpp"
+#include "util/ped.hpp"
 #include "views/view.hpp"
 
 namespace big
 {
-	static int range(int lower_bound, int upper_bound)
-	{
-		return std::rand() % (upper_bound - lower_bound + 1) + lower_bound;
-	}
-
 	void view::outfit_editor()
 	{
-		struct outfit_t
-		{
-			int id;
-			std::string label;
-			int drawable_id     = 0;
-			int drawable_id_max = 0;
-			int texture_id      = 0;
-			int texture_id_max  = 0;
-		};
-
-		static std::vector<outfit_t> components = {{0, "OUTFIT_HEAD"_T.data()},
-		    {1, "OUTFIT_BERD"_T.data()},
-		    {2, "OUTFIT_HAIR"_T.data()},
-		    {3, "OUTFIT_UPPR"_T.data()},
-		    {4, "OUTFIT_LOWR"_T.data()},
-		    {5, "OUTFIT_HAND"_T.data()},
-		    {6, "OUTFIT_FEET"_T.data()},
-		    {7, "OUTFIT_TEEF"_T.data()},
-		    {8, "OUTFIT_ACCS"_T.data()},
-		    {9, "OUTFIT_TASK"_T.data()},
-		    {10, "OUTFIT_DECL"_T.data()},
-		    {11, "OUTFIT_JBIB"_T.data()}};
-		static std::vector<outfit_t> props      = {{0, "OUTFIT_HEAD"_T.data()},
-		         {1, "OUTFIT_GLASSES"_T.data()},
-		         {2, "OUTFIT_EARS"_T.data()},
-		         {3, "OUTFIT_UNK1"_T.data()},
-		         {4, "OUTFIT_UNK2"_T.data()},
-		         {5, "OUTFIT_UNK3"_T.data()},
-		         {6, "OUTFIT_WATCH"_T.data()},
-		         {7, "OUTFIT_WRIST"_T.data()},
-		         {8, "OUTFIT_UNK4"_T.data()}};
+		static outfit::components_t components;
+		static outfit::props_t props;
 
 		g_fiber_pool->queue_job([] {
-			for (auto& item : components)
+			for (auto& item : components.items)
 			{
 				item.drawable_id     = PED::GET_PED_DRAWABLE_VARIATION(self::ped, item.id);
 				item.drawable_id_max = PED::GET_NUMBER_OF_PED_DRAWABLE_VARIATIONS(self::ped, item.id) - 1;
@@ -53,7 +21,7 @@ namespace big
 				item.texture_id_max = PED::GET_NUMBER_OF_PED_TEXTURE_VARIATIONS(self::ped, item.id, item.drawable_id) - 1;
 			}
 
-			for (auto& item : props)
+			for (auto& item : props.items)
 			{
 				item.drawable_id     = PED::GET_PED_PROP_INDEX(self::ped, item.id, 1);
 				item.drawable_id_max = PED::GET_NUMBER_OF_PED_PROP_DRAWABLE_VARIATIONS(self::ped, item.id) - 1;
@@ -64,12 +32,7 @@ namespace big
 		});
 
 		components::button("OUTFIT_RANDOM_COMPONENT"_T, [] {
-			for (auto& item : components)
-			{
-				int drawable_id = range(0, PED::GET_NUMBER_OF_PED_DRAWABLE_VARIATIONS(self::ped, item.id) - 1);
-				int texture_id = range(0, PED::GET_NUMBER_OF_PED_TEXTURE_VARIATIONS(self::ped, item.id, drawable_id) - 1);
-				PED::SET_PED_COMPONENT_VARIATION(self::ped, item.id, drawable_id, texture_id, PED::GET_PED_PALETTE_VARIATION(self::ped, item.id));
-			}
+			ped::set_ped_random_component_variation(self::ped);
 		});
 		ImGui::SameLine();
 
@@ -90,9 +53,9 @@ namespace big
 
 		components::button("EXPORT_TO_CLIPBOARD"_T, [] {
 			std::stringstream ss;
-			for (auto& item : components)
+			for (auto& item : components.items)
 				ss << item.id << " " << item.drawable_id << " " << item.texture_id << " ";
-			for (auto& item : props)
+			for (auto& item : props.items)
 				ss << item.id << " " << item.drawable_id << " " << item.texture_id << " ";
 			ImGui::SetClipboardText(ss.str().c_str());
 			g_notification_service->push("OUTFIT"_T.data(), "EXPORT_TO_CLIPBOARD"_T.data());
@@ -101,7 +64,7 @@ namespace big
 
 		components::button("IMPORT_FROM_CLIPBOARD"_T, [] {
 			std::stringstream ss(ImGui::GetClipboardText());
-			for (auto& item : components)
+			for (auto& item : components.items)
 			{
 				int id          = 0;
 				int drawable_id = 0;
@@ -111,7 +74,7 @@ namespace big
 				ss >> texture_id;
 				PED::SET_PED_COMPONENT_VARIATION(self::ped, id, drawable_id, texture_id, PED::GET_PED_PALETTE_VARIATION(self::ped, id));
 			}
-			for (auto& item : props)
+			for (auto& item : props.items)
 			{
 				int id          = 0;
 				int drawable_id = 0;
@@ -127,7 +90,7 @@ namespace big
 		});
 
 		ImGui::BeginGroup();
-		for (auto& item : components)
+		for (auto& item : components.items)
 		{
 			ImGui::SetNextItemWidth(60);
 			if (ImGui::InputInt(std::format("{} [0,{}]", item.label, item.drawable_id_max).c_str(), &item.drawable_id, 0))
@@ -142,7 +105,7 @@ namespace big
 		ImGui::SameLine();
 
 		ImGui::BeginGroup();
-		for (auto& item : components)
+		for (auto& item : components.items)
 		{
 			ImGui::SetNextItemWidth(60);
 			if (ImGui::InputInt(std::format("{} {} [0,{}]", item.label, "OUTFIT_TEX"_T, item.texture_id_max).c_str(), &item.texture_id, 0))
@@ -157,7 +120,7 @@ namespace big
 		ImGui::SameLine();
 
 		ImGui::BeginGroup();
-		for (auto& item : props)
+		for (auto& item : props.items)
 		{
 			ImGui::SetNextItemWidth(60);
 			if (ImGui::InputInt(std::format("{} [0,{}]", item.label, item.drawable_id_max).c_str(), &item.drawable_id, 0))
@@ -175,7 +138,7 @@ namespace big
 		ImGui::SameLine();
 
 		ImGui::BeginGroup();
-		for (auto& item : props)
+		for (auto& item : props.items)
 		{
 			ImGui::SetNextItemWidth(60);
 			if (ImGui::InputInt(std::format("{} {} [0,{}]", item.label, "OUTFIT_TEX"_T, item.texture_id_max).c_str(), &item.texture_id, 0))
@@ -206,7 +169,7 @@ namespace big
 			nlohmann::json j_components;
 			nlohmann::json j_props;
 
-			for (auto& item : components)
+			for (auto& item : components.items)
 			{
 				nlohmann::json tmp;
 				tmp["drawable_id"]                    = item.drawable_id;
@@ -214,7 +177,7 @@ namespace big
 				j_components[std::to_string(item.id)] = tmp;
 			}
 
-			for (auto& item : props)
+			for (auto& item : props.items)
 			{
 				nlohmann::json tmp;
 				tmp["drawable_id"]               = item.drawable_id;
