+++ src/views/settings/view_outfit_slots.cpp
@@ -22,40 +22,11 @@ namespace big
 			ImGui::SetNextItemWidth(300);
 			ImGui::InputText("OUTFIT_NAME"_T.data(), outfit::get_slot_name_address(slot), 16);
 
-			struct outfit_t
-			{
-				int id;
-				std::string label;
-				int drawable_id     = 0;
-				int drawable_id_max = 0;
-				int texture_id      = 0;
-				int texture_id_max  = 0;
-			};
-
-			static std::vector<outfit_t> components = {{0, "OUTFIT_HEAD"_T.data()},
-			    {1, "OUTFIT_BERD"_T.data()},
-			    {2, "OUTFIT_HAIR"_T.data()},
-			    {3, "OUTFIT_UPPR"_T.data()},
-			    {4, "OUTFIT_LOWR"_T.data()},
-			    {5, "OUTFIT_HAND"_T.data()},
-			    {6, "OUTFIT_FEET"_T.data()},
-			    {7, "OUTFIT_TEEF"_T.data()},
-			    {8, "OUTFIT_ACCS"_T.data()},
-			    {9, "OUTFIT_TASK"_T.data()},
-			    {10, "OUTFIT_DECL"_T.data()},
-			    {11, "OUTFIT_JBIB"_T.data()}};
-			static std::vector<outfit_t> props      = {{0, "OUTFIT_HEAD"_T.data()},
-			         {1, "OUTFIT_GLASSES"_T.data()},
-			         {2, "OUTFIT_EARS"_T.data()},
-			         {3, "OUTFIT_UNK1"_T.data()},
-			         {4, "OUTFIT_UNK2"_T.data()},
-			         {5, "OUTFIT_UNK3"_T.data()},
-			         {6, "OUTFIT_WATCH"_T.data()},
-			         {7, "OUTFIT_WRIST"_T.data()},
-			         {8, "OUTFIT_UNK4"_T.data()}};
+			static outfit::components_t components;
+			static outfit::props_t props;
 
 			g_fiber_pool->queue_job([] {
-				for (auto& item : components)
+				for (auto& item : components.items)
 				{
 					item.drawable_id     = *outfit::get_component_drawable_id_address(slot, item.id);
 					item.drawable_id_max = PED::GET_NUMBER_OF_PED_DRAWABLE_VARIATIONS(self::ped, item.id) - 1;
@@ -64,7 +35,7 @@ namespace big
 					item.texture_id_max = PED::GET_NUMBER_OF_PED_TEXTURE_VARIATIONS(self::ped, item.id, item.drawable_id) - 1;
 				}
 
-				for (auto& item : props)
+				for (auto& item : props.items)
 				{
 					item.drawable_id     = *outfit::get_prop_drawable_id_address(slot, item.id);
 					item.drawable_id_max = PED::GET_NUMBER_OF_PED_PROP_DRAWABLE_VARIATIONS(self::ped, item.id) - 1;
@@ -76,9 +47,9 @@ namespace big
 
 			components::button("EXPORT_TO_CLIPBOARD"_T, [] {
 				std::stringstream ss;
-				for (auto& item : components)
+				for (auto& item : components.items)
 					ss << item.id << " " << item.drawable_id << " " << item.texture_id << " ";
-				for (auto& item : props)
+				for (auto& item : props.items)
 					ss << item.id << " " << item.drawable_id << " " << item.texture_id << " ";
 				ImGui::SetClipboardText(ss.str().c_str());
 				g_notification_service->push("OUTFIT"_T.data(), "EXPORT_TO_CLIPBOARD"_T.data());
@@ -87,7 +58,7 @@ namespace big
 
 			components::button("IMPORT_FROM_CLIPBOARD"_T, [] {
 				std::stringstream ss(ImGui::GetClipboardText());
-				for (auto& item : components)
+				for (auto& item : components.items)
 				{
 					int id          = 0;
 					int drawable_id = 0;
@@ -98,7 +69,7 @@ namespace big
 					*outfit::get_component_drawable_id_address(slot, id) = drawable_id;
 					*outfit::get_component_texture_id_address(slot, id)  = texture_id;
 				}
-				for (auto& item : props)
+				for (auto& item : props.items)
 				{
 					int id          = 0;
 					int drawable_id = 0;
@@ -112,7 +83,7 @@ namespace big
 			});
 
 			ImGui::BeginGroup();
-			for (auto& item : components)
+			for (auto& item : components.items)
 			{
 				ImGui::SetNextItemWidth(60);
 				ImGui::InputInt(std::format("{} [0,{}]", item.label, item.drawable_id_max).c_str(), outfit::get_component_drawable_id_address(slot, item.id), 0);
@@ -122,7 +93,7 @@ namespace big
 			ImGui::SameLine();
 
 			ImGui::BeginGroup();
-			for (auto& item : components)
+			for (auto& item : components.items)
 			{
 				ImGui::SetNextItemWidth(60);
 				ImGui::InputInt(std::format("{} {} [0,{}]", item.label, "OUTFIT_TEX"_T, item.texture_id_max).c_str(), outfit::get_component_texture_id_address(slot, item.id), 0);
@@ -132,7 +103,7 @@ namespace big
 			ImGui::SameLine();
 
 			ImGui::BeginGroup();
-			for (auto& item : props)
+			for (auto& item : props.items)
 			{
 				ImGui::SetNextItemWidth(60);
 				ImGui::InputInt(std::format("{} [0,{}]", item.label, item.drawable_id_max).c_str(), outfit::get_prop_drawable_id_address(slot, item.id), 0);
@@ -142,7 +113,7 @@ namespace big
 			ImGui::SameLine();
 
 			ImGui::BeginGroup();
-			for (auto& item : props)
+			for (auto& item : props.items)
 			{
 				ImGui::SetNextItemWidth(60);
 				ImGui::InputInt(std::format("{} {} [0,{}]", item.label, "OUTFIT_TEX"_T, item.texture_id_max).c_str(), outfit::get_prop_texture_id_address(slot, item.id), 0);
