+++ src/views/world/view_spawn_ped.cpp
@@ -134,7 +134,7 @@ namespace big
 		PED::SET_PED_ID_RANGE(ped, 200.0f);
 		PED::SET_PED_FIRING_PATTERN(ped, RAGE_JOAAT("FIRING_PATTERN_FULL_AUTO"));
 		PED::SET_PED_SHOOT_RATE(ped, 150);
-		PED::SET_PED_RANDOM_COMPONENT_VARIATION(ped, 0);
+		ped::set_ped_random_component_variation(ped);
 
 		if (is_bodyguard)
 		{
@@ -617,7 +617,7 @@ namespace big
 					return;
 				}
 
-				PED::SET_PED_RANDOM_COMPONENT_VARIATION(self::ped, 0);
+				ped::set_ped_random_component_variation(self::ped);
 			}
 		});
